ci(desktop): isolate Keychain contract checks
